requirements.txt
# ─── Core Deep Learning ───────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ─── Face Detection & Embedding ───────────────────────────────────────────────
facenet-pytorch>=2.5.3       # MTCNN + InceptionResnetV1 (ArcFace-style)

# ─── Computer Vision ──────────────────────────────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0        # Data augmentation for training

# ─── Numerical & Data ─────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# ─── Vector Search ────────────────────────────────────────────────────────────
faiss-cpu>=1.7.4             # Use faiss-gpu if CUDA available and installed separately

# ─── Explainability ───────────────────────────────────────────────────────────
captum>=0.7.0                # Grad-CAM and other attribution methods

# ─── API Framework ────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6      # File uploads
aiofiles>=23.0.0             # Async file handling

# ─── Data Validation ──────────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ─── Database ─────────────────────────────────────────────────────────────────
# SQLite is built-in to Python; no extra package needed

# ─── Training Utilities ───────────────────────────────────────────────────────
tqdm>=4.66.0                 # Progress bars
tensorboard>=2.14.0          # Training curve visualization
scikit-learn>=1.3.0          # Metrics, preprocessing
matplotlib>=3.7.0            # Plotting

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
httpx>=0.25.0                # Async FastAPI test client
